Cleaned up CPU tests. Separated the test file into a header and source file. CPUTestFixture fixture class renamed to CpuTest for easier usage.
